01.03

+++ board-management-web/src/main/java/kr/co/ymtech/bm/service/BoardService.java
@@ -102,7 +102,7 @@ public class BoardService implements IBoardService {
 	 */
 	@Override
 	public Integer saveBoard(BoardDTO board) {
-		
+
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 
 		String originalFileName = null;
@@ -174,44 +174,49 @@ public class BoardService implements IBoardService {
 		String deleteFileId = null;
 		List<String> deleteFiles = board.getDeleteFiles();
 		List<FileVO> boardFiles = new ArrayList<FileVO>();
-
-		// dto -> vo 변환
 		BoardVO vo = new BoardVO();
-		vo.setIndex(board.getIndex());
-		vo.setTitle(board.getTitle());
-		vo.setText(board.getText());
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 
-		// 추가된 파일의 크기에 따라 반복하여 파일을 저장
-		for (MultipartFile file : board.getAddFiles()) {
-			originalFileName = file.getOriginalFilename();
-			uniqueID = UUID.randomUUID().toString();
-			filePath = Paths.get(PathConfig.getImagePath()).resolve(uniqueID + "_" + originalFileName).normalize()
-					.toString();
+		// 로그인한 유저 아이디가 작성자 아이디와 같거나 "admin"이면 동작
+		if (board.getUserId().equals(auth.getName()) || "admin".equals(auth.getName())) {
 
-			boardFile = new FileVO();
-			boardFile.setFileId(uniqueID);
-			boardFile.setFilePath(PathConfig.getImagePath());
-			boardFile.setFileName(originalFileName);
-			boardFile.setFileSize(file.getSize());
+			// dto -> vo 변환
+			vo.setIndex(board.getIndex());
+			vo.setTitle(board.getTitle());
+			vo.setText(board.getText());
 
-			boardFiles.add(boardFile);
+			// 추가된 파일의 크기에 따라 반복하여 파일을 저장
+			for (MultipartFile file : board.getAddFiles()) {
+				originalFileName = file.getOriginalFilename();
+				uniqueID = UUID.randomUUID().toString();
+				filePath = Paths.get(PathConfig.getImagePath()).resolve(uniqueID + "_" + originalFileName).normalize()
+						.toString();
 
-			// 추가된 파일을 지정된 경로의 폴더에 저장
-			try (InputStream input = file.getInputStream(); OutputStream output = new FileOutputStream(filePath)) {
-				IOUtils.copy(input, output);
-			} catch (IOException e) {
-				System.out.println("파일 저장 실패");
+				boardFile = new FileVO();
+				boardFile.setFileId(uniqueID);
+				boardFile.setFilePath(PathConfig.getImagePath());
+				boardFile.setFileName(originalFileName);
+				boardFile.setFileSize(file.getSize());
+
+				boardFiles.add(boardFile);
+
+				// 추가된 파일을 지정된 경로의 폴더에 저장
+				try (InputStream input = file.getInputStream(); OutputStream output = new FileOutputStream(filePath)) {
+					IOUtils.copy(input, output);
+				} catch (IOException e) {
+					System.out.println("파일 저장 실패");
+				}
 			}
-		}
 
-		// 삭제한 파일 리스트를 반복하며 파일 삭제
-		for (String deleteFileName : deleteFiles) {
-			deleteFilePath = Paths.get(PathConfig.getImagePath()).resolve(deleteFileName).normalize().toString();
-			deleteFileId = deleteFileName.substring(0, deleteFileName.indexOf("_"));
-			deleteFile = new File(deleteFilePath);
-			deleteFile.delete();
+			// 삭제한 파일 리스트를 반복하며 파일 삭제
+			for (String deleteFileName : deleteFiles) {
+				deleteFilePath = Paths.get(PathConfig.getImagePath()).resolve(deleteFileName).normalize().toString();
+				deleteFileId = deleteFileName.substring(0, deleteFileName.indexOf("_"));
+				deleteFile = new File(deleteFilePath);
+				deleteFile.delete();
 
-			boardRepository.deleteFile(board.getIndex(), deleteFileId);
+				boardRepository.deleteFile(board.getIndex(), deleteFileId);
+			}
 		}
 
 		return boardRepository.updateBoard(vo, boardFiles);
@@ -228,25 +233,32 @@ public class BoardService implements IBoardService {
 	 * @since 2023. 10. 26.
 	 */
 	@Override
-	public Integer deleteBoard(Integer index) {
+	public Integer deleteBoard(Integer index, String userId) {
 
 		String filePath = null;
 		List<FileVO> files = boardRepository.files(index);
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 
-		// 삭제하려는 게시글에 업로드된 파일을 지정된 경로의 폴더에서 삭제
-		for (FileVO file : files) {
-			filePath = Paths.get(PathConfig.getImagePath()).resolve(file.getFileId() + "_" + file.getFileName())
-					.normalize().toString();
+		// 로그인한 유저 아이디가 작성자 아이디와 같거나 "admin"이면 동작
+		if (userId.equals(auth.getName()) || "admin".equals(auth.getName())) {
 
-			File deleteFile = new File(filePath);
-			deleteFile.delete();
+			// 삭제하려는 게시글에 업로드된 파일을 지정된 경로의 폴더에서 삭제
+			for (FileVO file : files) {
+				filePath = Paths.get(PathConfig.getImagePath()).resolve(file.getFileId() + "_" + file.getFileName())
+						.normalize().toString();
+
+				File deleteFile = new File(filePath);
+				deleteFile.delete();
+			}
+
+			boardRepository.deleteFiles(index);
+
+			commentRepository.deleteAllComment(index);
+
+			return boardRepository.deleteBoard(index);
 		}
 
-		boardRepository.deleteFiles(index);
-
-		commentRepository.deleteAllComment(index);
-
-		return boardRepository.deleteBoard(index);
+		return null;
 	}
 
 	/**
@@ -265,6 +277,8 @@ public class BoardService implements IBoardService {
 		List<FileVO> fv = boardRepository.files(index);
 		BoardVO vo = boardRepository.searchByIndex(index);
 		BoardGetDTO dto = new BoardGetDTO();
+		Integer likeCount = boardRepository.boardLikeCount(index);
+		Integer userLike = boardRepository.checkUserBoardLike(index);
 
 		// vo -> dto 변환
 		dto.setIndex(vo.getIndex());
@@ -273,28 +287,28 @@ public class BoardService implements IBoardService {
 		dto.setUserId(vo.getUserId());
 		dto.setCategory(vo.getCategory());
 		dto.setCreateDate(new Long(vo.getCreateDate()));
-		dto.setLikeCount(vo.getLikeCount());
+		dto.setLikeCount(likeCount);
+		dto.setUserLike(userLike);
 		dto.setFile(fv);
 		return dto;
 	}
 
 	/**
-	 * @Method boardLikeCount 해당 게시글의 추천 수를 반환하는 메소드
+	 * @Method updateBoardLike 게시글 추천 수를 업데이트하는 메소드
 	 *
-	 * @see kr.co.ymtech.bm.repository.IBoardRepository#boardLikeCount(java.lang.Integer,
-	 *      java.lang.Integer)
+	 * @see kr.co.ymtech.bm.service.IBoardService#updateBoardLike(java.lang.Integer)
 	 *
-	 * @param index     해당 게시글 번호
-	 * @param likeCount 해당 게시글 추천 수
+	 * @param index 해당 게시글 번호
 	 * 
-	 * @return boardRepository의 boardLikeCount메소드 실행
+	 * @return 게시글 추천 수를 업데이트하는 Repository 함수 실행
 	 *
 	 * @author 황상필
-	 * @since 2023. 11. 03.
+	 * @since 2024. 01. 02.
 	 */
 	@Override
-	public Integer boardLikeCount(Integer index, Integer likeCount) {
-		return boardRepository.boardLikeCount(index, likeCount);
+	public Integer updateBoardLike(Integer index) {
+		boardRepository.updateBoardLike(index);
+		return boardRepository.bestBoardLike(index);
 	}
 
 	/**
